+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Before releasing:
 - Added a `toggle` method to `AdiDigitalOut` to toggle between level outputs.
 - Added a `SerialPort::set_baud_rate` method for the adjusting baudrate of a generic serial smartport after initialization. (#217)
 - Added fields containing relevant failure information to several error types (#221) (**Breaking Change**)
+- Added support for the power button in the `Controller` API. (#231) (**Breaking Change**)
 
 ### Fixed